+++ bin/faucet/src/faucet/mod.rs
@@ -8,7 +8,7 @@ use miden_client::{
     builder::ClientBuilder,
     crypto::RpoRandomCoin,
     keystore::FilesystemKeyStore,
-    note::{Note, NoteError, create_p2id_note},
+    note::{NoteError, NoteId, create_p2id_note},
     rpc::Endpoint,
     transaction::{
         LocalTransactionProver, OutputNote, TransactionId, TransactionProver,
@@ -20,12 +20,14 @@ use rand::{rng, rngs::StdRng};
 use serde::Serialize;
 use tokio::sync::mpsc::Receiver;
 use tracing::{info, instrument, warn};
-use updates::{ClientUpdater, MintUpdate, ResponseSender};
+use updates::{ClientUpdater, MintUpdate};
 use url::Url;
 
 use crate::types::{AssetAmount, NoteType};
 
 mod updates;
+
+pub use updates::MintResponseSender;
 
 // FAUCET CLIENT
 // ================================================================================================
@@ -157,14 +159,14 @@ impl Faucet {
     /// error.
     pub async fn run(
         mut self,
-        mut requests: Receiver<(MintRequest, ResponseSender)>,
+        mut requests: Receiver<(MintRequest, MintResponseSender)>,
     ) -> anyhow::Result<()> {
         let mut buffer = Vec::new();
-        let limit = 100; // we could include 256 notes per tx, but requests channel is limited to 100 atm
+        let limit = 256; // also limited by the queue size `REQUESTS_QUEUE_SIZE`
 
         while requests.recv_many(&mut buffer, limit).await > 0 {
             // Skip requests where the user no longer cares about the result.
-            let (requests, response_senders): (Vec<MintRequest>, Vec<ResponseSender>) = buffer
+            let (requests, response_senders): (Vec<MintRequest>, Vec<MintResponseSender>) = buffer
                 .drain(..)
                 .filter(|(request, response_sender)| {
                     if response_sender.is_closed() {
@@ -176,12 +178,17 @@ impl Faucet {
                 })
                 .unzip();
 
-            let updater = ClientUpdater::new(response_senders, self.id.network_id);
+            let updater = ClientUpdater::new(response_senders);
 
-            let (notes, tx_id) = self.handle_request_batch(&requests, &updater).await?;
+            let (tx_id, note_ids) = self.handle_request_batch(&requests, &updater).await?;
 
-            let block_height = self.client.get_sync_height().await?;
-            updater.send_notes(block_height, &notes, tx_id).await;
+            for note_id in note_ids {
+                updater
+                    .send_updates(MintUpdate::Minted(note_id, tx_id, self.id.network_id))
+                    .await;
+            }
+
+            self.client.sync_state().await?;
         }
 
         tracing::info!("Request stream closed, shutting down minter");
@@ -198,14 +205,13 @@ impl Faucet {
         &mut self,
         requests: &[MintRequest],
         updater: &ClientUpdater,
-    ) -> Result<(Vec<Note>, TransactionId), ClientError> {
+    ) -> Result<(TransactionId, Vec<NoteId>), ClientError> {
         let mut rng = get_rpo_random_coin(&mut rng());
 
         // Build the notes
         let notes = build_p2id_notes(self.id, self.decimals, requests, &mut rng)?;
-        let output_notes: Vec<OutputNote> =
-            notes.clone().into_iter().map(OutputNote::Full).collect();
-        let tx = TransactionRequestBuilder::new().own_output_notes(output_notes).build()?;
+        let note_ids = notes.iter().map(OutputNote::id).collect();
+        let tx = TransactionRequestBuilder::new().own_output_notes(notes).build()?;
         updater.send_updates(MintUpdate::Built).await;
 
         // Execute the transaction
@@ -225,7 +231,7 @@ impl Faucet {
         }
         updater.send_updates(MintUpdate::Submitted).await;
 
-        Ok((notes, tx_id))
+        Ok((tx_id, note_ids))
     }
 
     /// Returns the id of the faucet account.
@@ -247,7 +253,7 @@ fn build_p2id_notes(
     decimals: u8,
     requests: &[MintRequest],
     rng: &mut RpoRandomCoin,
-) -> Result<Vec<Note>, NoteError> {
+) -> Result<Vec<OutputNote>, NoteError> {
     // If building a note fails, we discard the whole batch. Should never happen, since account
     // ids are validated on the request level.
     let mut notes = Vec::new();
@@ -263,7 +269,7 @@ fn build_p2id_notes(
                 Felt::default(),
                 rng,
             ).inspect_err(|err| tracing::error!(request.account_id=%request.account_id, ?err, "failed to build note"))?;
-        notes.push(note);
+        notes.push(OutputNote::Full(note));
     }
     Ok(notes)
 }
@@ -288,7 +294,7 @@ mod tests {
     use url::Url;
 
     use super::*;
-    use crate::{stub_rpc_api::serve_stub, types::AssetOptions};
+    use crate::{testing::stub_rpc_api::serve_stub, types::AssetOptions};
 
     // This test ensures that the faucet can create a transaction that outputs a batch of notes.
     #[allow(clippy::cast_sign_loss)]
@@ -359,8 +365,8 @@ mod tests {
             })
             .collect::<Vec<_>>();
 
-        let (notes, tx_id) = faucet
-            .handle_request_batch(&requests, &ClientUpdater::new(vec![], NetworkId::Testnet))
+        let (tx_id, note_ids) = faucet
+            .handle_request_batch(&requests, &ClientUpdater::new(vec![]))
             .await
             .unwrap();
 
@@ -373,6 +379,9 @@ mod tests {
             .unwrap()
             .clone();
         assert_eq!(tx.details.output_notes.num_notes(), num_requests as usize);
-        assert_eq!(notes.len(), num_requests as usize);
+        assert_eq!(
+            tx.details.output_notes.iter().map(OutputNote::id).collect::<Vec<_>>(),
+            note_ids
+        );
     }
 }