+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - Integrated miden-client (#11).
+- Added `/get_note` endpoint (#19).
 - Redesigned the home frontend (#20).
 
 ## 0.10.0 (2025-07-10)